requirements.txt
asgiref==3.6.0
autopep8==2.0.1
# backports.zoneinfo==0.2.1
Django==4.1.4
django-formtools==2.4
isort==5.11.4
pycodestyle==2.10.0
sqlparse==0.4.3
toml==0.10.2
Pillow~=9.3.0